+++ CHANGES.md
@@ -1,10 +1,12 @@
 # Change Log
 
-### ? - ?
+
+### Not Released Yet
 
 ##### Breaking Changes :mega:
 
 - Renamed `shouldContentContinueUpdating` to `getMightHaveLatentChildren` and `setContentShouldContinueUpdating` to `setMightHaveLatentChildren` on the `Tile` class.
+- `LoadedRasterOverlayImage` now has a single `errorList` property instead of separate `errors` and `warnings` properties.
 
 ##### Additions :tada:
 
@@ -16,6 +18,8 @@
 - Fixed a bug that prevented use of `Tileset` with a nullptr `IPrepareRendererResources`.
 - Fixed a bug in `IntersectionTests::rayOBBParametric` that could cause incorrect results for some oriented bounding boxes.
 - `GltfUtilities::intersectRayGltfModel` now reports a warning when given a model it can't compute the intersection with because it uses required extensions that are not supported.
+- Errors while loading raster overlays are now logged. Previously, they were silently ignored in many cases.
+- A raster overlay image failing to load will no longer completely prevent the geometry tile to which it is attached from rendering. Instead, once the raster overlay fails, the geometry tile will be shown without the raster overlay.
 
 ### v0.39.0 - 2024-09-02